# api-gateway/cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing users, tweets, comments, likes, and direct messages.",
        "title": "API Gateway",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:5050",
    "basePath": "/",
    "paths": {
...
        "/directs/{id}": {
            "get": {
                "description": "Get a specific direct message by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "direct_messages"
                ],
                "summary": "Get a direct message by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Direct Message ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api-gateway_protos_direct-proto.GetDirectMessageByIDResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers_direct-handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers_direct-handlers.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a specific direct message by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "direct_messages"
                ],
                "summary": "Delete a direct message",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Direct Message ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api-gateway_protos_direct-proto.DeleteDirectMessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers_direct-handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers_direct-handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/directs/{sender_id}/{receiver_id}": {
            "get": {
                "description": "Get all direct messages between two users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "direct_messages"
                ],
                "summary": "Get direct messages",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sender ID",
                        "name": "sender_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Receiver ID",
                        "name": "receiver_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api-gateway_protos_direct-proto.GetDirectMessagesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers_direct-handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers_direct-handlers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new direct message between users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "direct_messages"
                ],
                "summary": "Create a direct message",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sender ID",
                        "name": "sender_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Receiver ID",
                        "name": "receiver_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Message text",
                        "name": "text",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Media files",
                        "name": "media",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api-gateway_protos_direct-proto.CreateDirectMessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers_direct-handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers_direct-handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api-gateway_protos_direct-proto.CreateDirectMessageResponse": {
            "type": "object",
            "properties": {
                "direct_message": {
                    "$ref": "#/definitions/api-gateway_protos_direct-proto.DirectMessage"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "api-gateway_protos_direct-proto.DeleteDirectMessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "api-gateway_protos_direct-proto.DirectMessage": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "media": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "receiver_id": {
                    "type": "integer"
                },
                "sender_id": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                },
                "tweet_id": {
                    "type": "integer"
                }
            }
        },
        "api-gateway_protos_direct-proto.GetDirectMessageByIDResponse": {
            "type": "object",
            "properties": {
                "direct_message": {
                    "$ref": "#/definitions/api-gateway_protos_direct-proto.DirectMessage"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "api-gateway_protos_direct-proto.GetDirectMessagesResponse": {
            "type": "object",
            "properties": {
                "direct_messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api-gateway_protos_direct-proto.DirectMessage"
                    }
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "internal_handlers_direct-handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
    }
}